refactor: split the monolith into config/debuglog/win32utils/worker modules
Break the single ~4.1k-line claude_overlay.py into focused modules with a clean,
acyclic dependency graph. This is a move-only refactor: the extracted code is
byte-identical to the original (the claude_overlay.py diff is +6/-1023 - only the
import head changed; the ~3k-line Overlay class is untouched).

  config.py          constants, theme palette, SYSTEM_APPEND, __version__ (leaf)
  debuglog.py        dbg() + the worker->UI queue tap (leaf)
  win32utils.py      DPI / AppUserModelID / window-region / monitor enum (-> config)
  worker.py          ClaudeWorker + the SDK imports + the anyio no-console patch
  claude_overlay.py  the Overlay UI class + __main__ (imports the above)

Also:
- Make STRICT_MCP_CONFIG env-driven (CLAUDE_OVERLAY_STRICT_MCP, default True) so a
  release no longer has to hand-toggle the constant; a machine sets the env var to
  inherit its own MCP servers.
- Add a pytest suite under tests/ (93 tests) and a Windows CI workflow that
  byte-compiles every module, import-smoke-tests the app, and runs the suite.

Verified: py_compile (all 5 modules), `import claude_overlay`, 93 passed / 2 xfailed,
uniform CRLF, moved-name coverage grep clean. No behavior change.
