+++ .github/workflows/tests.yml
@@ -1,0 +1,34 @@
+name: tests
+
+# Byte-compile every module, import the app headlessly, and run the unit suite on a
+# real Windows runner (the app is Windows-only: win32utils calls Win32 via ctypes, and
+# the import smoke pulls in tkinter). No `claude` login is needed — nothing connects.
+
+on:
+  push:
+  pull_request:
+
+permissions:
+  contents: read
+
+jobs:
+  test:
+    runs-on: windows-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.x"
+
+      - name: Install dependencies
+        run: python -m pip install -r requirements-dev.txt
+
+      - name: Byte-compile all modules
+        run: python -m py_compile config.py debuglog.py win32utils.py worker.py claude_overlay.py
+
+      - name: Import smoke test (no window, no connect)
+        run: python -c "import claude_overlay; print('import OK', claude_overlay.__version__)"
+
+      - name: Run unit tests
+        run: python -m pytest
