+++ claude_overlay.py
@@ -30,1029 +30,12 @@ from tkinter import font as tkfont
 
 from PIL import Image, ImageGrab, ImageDraw, ImageChops, ImageFilter, ImageTk
 
-# Make sure both common `claude` install locations are on PATH, in case it was just
-# installed this session (PATH not yet refreshed): the native installer drops it in
-# %USERPROFILE%\.local\bin, and a global npm install in %APPDATA%\npm.
-os.environ["PATH"] = os.pathsep.join(filter(None, [
-    os.path.join(os.environ.get("USERPROFILE", ""), ".local", "bin"),
-    os.path.join(os.environ.get("APPDATA", ""), "npm"),
-    os.environ.get("PATH", ""),
-]))
-
-def _env_int(name: str, default: int, min_value: int, max_value: int) -> int:
-    try:
-        val = int(os.environ.get(name, ""))
-    except Exception:
-        return default
-    return max(min_value, min(max_value, val))
-
-
-# Spawn the `claude` CLI subprocess with no console window. Without this, running
-# under pythonw (no console) makes Windows pop a CMD window for the console-mode CLI.
-# Best-effort: if a future anyio drops/renames open_process, degrade gracefully
-# (worst case a CMD window flashes) rather than crash on import.
-try:
-    import anyio as _anyio  # noqa: E402
-    if sys.platform == "win32":
-        _CREATE_NO_WINDOW = 0x08000000
-        _orig_open_process = _anyio.open_process
-
-        async def _open_process_no_window(*args, **kwargs):
-            kwargs["creationflags"] = kwargs.get("creationflags", 0) | _CREATE_NO_WINDOW
-            return await _orig_open_process(*args, **kwargs)
-
-        _anyio.open_process = _open_process_no_window
-except Exception:
-    pass
-
-from claude_agent_sdk import (  # noqa: E402
-    ClaudeSDKClient, ClaudeAgentOptions, AssistantMessage, TextBlock,
-    ToolUseBlock, ResultMessage, StreamEvent, PermissionResultAllow,
-)
-# Error types used to decide when the transport is broken and we should reconnect.
-# Imported defensively: older/newer SDKs may not export all of them.
-try:
-    from claude_agent_sdk import (  # noqa: E402
-        ClaudeSDKError, CLIConnectionError, CLIJSONDecodeError, ProcessError,
-    )
-except Exception:  # pragma: no cover
-    class ClaudeSDKError(Exception): ...
-    class CLIConnectionError(ClaudeSDKError): ...
-    class CLIJSONDecodeError(ClaudeSDKError): ...
-    class ProcessError(ClaudeSDKError): ...
-
-__version__ = "1.9.0"
-
-# ───────────────────────────── configuration ──────────────────────────────
-WORKING_DIR = str(Path.home())
-# NOTE: the Agent SDK's model=None does NOT follow the CLI's interactive default
-# (which is opus-4-8); SDK 0.2.87 resolves None → opus-4-7. So pin the ID explicitly.
-# Default to the standard 200K-context Opus: overlay chats never approach 200K, so the
-# "[1m]" 1M-context variant only buys latency for context we never use. The 1M variant
-# stays one click away in the MODELS switcher.
-MODEL = "claude-opus-4-8"
-MODELS = [("Opus 4.8", "claude-opus-4-8"), ("Opus 4.8 (1M)", "claude-opus-4-8[1m]"),
-          ("Sonnet", "sonnet"), ("Haiku", "haiku")]  # click the statusline to switch
-PERMISSION_MODE = "bypassPermissions"
-STRICT_MCP_CONFIG = True          # do NOT inherit the user's ~/.claude MCP servers. The overlay
-                                  # is a lightweight screen-chat that only needs the core Claude
-                                  # Code tools; inheriting every MCP server the user has configured
-                                  # (Atlassian, Figma, M365, …) injects their tool schemas into the
-                                  # context — easily 50-70K+ tokens, a third of a 200K window, gone
-                                  # before you type. Flip to False to expose your MCP tools here.
-SKILLS = "all"                    # which Agent SDK skills to enable in the overlay. Default None
-                                  # means the overlay discovers NO skills (the SDK only wires up
-                                  # skill discovery when this is set). A list enables ONLY those
-                                  # skills by name — lean, just their description lands in context.
-                                  # "all" enables every discovered skill (heavier: every skill's
-                                  # description is injected). Setting this also makes the CLI load
-                                  # ~/.claude settings (setting_sources defaults to user+project),
-                                  # but NOT MCP servers (STRICT_MCP_CONFIG still blocks those).
-                                  # None → skills off entirely.
-AUTO_SCREENSHOT_DEFAULT = True
-SHOW_IN_SCREEN_SHARE_DEFAULT = False  # False (default) = the overlay is excluded from screen
-                                  # captures at the OS/DWM level (WDA_EXCLUDEFROMCAPTURE): it
-                                  # stays visible to YOU but is omitted from Teams/Zoom/Meet/OBS
-                                  # screen shares, PrintScreen, and our own screenshots — private.
-                                  # True = the overlay shows up in screen shares. Flip it live via
-                                  # the status-bar "shareable" toggle; no restart needed.
-HIDE_SCREENSHOT_TOOL = True       # hide the noisy "⚙ Read …shot_*.png" lines every turn
-HOTKEY = "ctrl+alt+space"
-THEME = "light"                  # "light" (Claude paper) or "dark" (warm dark)
-WINDOW_ALPHA = 1.0
-CORNER_RADIUS = 18
-TASKBAR_BUTTON = True            # show a real, clickable Windows taskbar button (with the
-                                 # Clawd icon), like a normal app — alt-tab target, click to
-                                 # focus/raise, see at a glance that it's running. The frameless
-                                 # (overrideredirect) window gets NO taskbar button by default;
-                                 # this forces one via WS_EX_APPWINDOW. False → the pure
-                                 # no-taskbar floating overlay (original behaviour).
-APP_ICON = "claude_overlay_2.ico"  # window + taskbar icon (Clawd). Path is relative to this
-                                 # script (or absolute). "" → no custom icon (Tk default).
-APP_ID = "shengyanlin.claude-overlay"  # explicit Windows AppUserModelID. Without it a pythonw
-                                 # app shows pythonw's icon in the taskbar and groups with other
-                                 # Python apps; setting it makes the taskbar use APP_ICON instead.
-ORB_SIZE = 56                    # diameter (logical px) of the collapsed Claude orb
-ORB_IMAGE = "claude_overlay_2.png"  # collapsed-orb artwork. "" → procedural glossy
-                                 # terracotta sphere (original look). A path (relative
-                                 # to this script, or absolute) to a PNG/ICO renders that
-                                 # image instead: it's auto-scaled + centred so the whole
-                                 # opaque shape fits inside the circular orb. RGBA with a
-                                 # transparent background works best.
-ORB_IMAGE_MARGIN = 0.04          # fraction of the radius kept clear around the artwork
-                                 # (0 = touches the circle edge; 0.04 = tiny breathing room)
-ORB_FLOAT = True                 # True + an ORB_IMAGE → the collapsed orb is clipped to the
-                                 # artwork's own silhouette (a free-floating pixel sprite, no
-                                 # circular frame; clicks outside the shape pass through).
-                                 # False → the classic circular orb. Ignored without artwork.
-ORB_ALPHA_THRESHOLD = 110        # pixels at/above this alpha (0-255) count as "solid" when
-                                 # building the silhouette — higher = tighter, crisper edge
-# Fonts. Noto Sans/Serif TC cover Chinese + English in one family (closest free
-# stand-in for Claude's proprietary Styrene/Copernicus). First available wins.
-FONT_SANS = ["Noto Sans TC", "Inter", "Segoe UI Variable Text", "Segoe UI"]
-FONT_SERIF = ["Noto Serif TC", "Georgia", "Cambria"]   # the "Claude" wordmark
-FONT_MONO = ["Consolas", "Cascadia Mono", "Courier New"]
-SHOT_DIR = Path(os.environ.get("TEMP", str(Path.home()))) / "claude_overlay_shots"
-KEEP_SHOTS = 24                  # retain a few captures worth (one file per monitor)
-SHOT_MAX_EDGE = 1568             # downscale captures to this long edge before sending.
-SHOT_FORMAT = os.environ.get("CLAUDE_OVERLAY_SHOT_FORMAT", "auto").strip().lower()
-                                 # "auto" saves PNG + JPEG and keeps the smaller payload;
-                                 # "png" preserves old behavior; "jpeg" favors upload speed.
-SHOT_JPEG_QUALITY = _env_int("CLAUDE_OVERLAY_SHOT_JPEG_QUALITY", 82, 50, 95)
-                                 # Claude downsamples larger images internally anyway, so
-                                 # bigger files only cost upload time + vision tokens.
-IMAGE_INPUT = "inline"           # "inline" → attach screenshots as base64 image blocks
-                                 # (no per-turn Read round-trip); "read" → legacy path:
-                                 # save PNG + ask Claude to Read it. Flip to "read" if a
-                                 # future CLI rejects inline images.
-PRECAPTURE_ON_TYPING = True      # grab the screen ~as you type (off the send path) so
-                                 # send latency excludes the capture.
-PRECAPTURE_MAX_AGE = 6.0         # seconds a pre-captured frame stays reusable; older than
-                                 # this at send time → re-grab fresh (bounds staleness).
-MAX_BUFFER_SIZE = 64 * 1024 * 1024   # the SDK aborts a turn with CLIJSONDecodeError when a
-                                 # single stream-json line exceeds this (default 1MB). Inline
-                                 # screenshots (base64, ×monitors) blow past 1MB easily and
-                                 # used to crash the worker — 64MB gives huge headroom.
-# A hang is NOT an exception, so the reconnect / bounded-restart guards (which only fire
-# on a raised error) can't preempt an SDK call that never resolves — a wedged transport
-# (broken corporate TLS, half-open socket, CLI waiting on a prompt with no TTY) would pin
-# the worker forever. Bound every SDK lifecycle call so a hang degrades to a clean
-# reconnect instead of a permanent freeze.
-CONNECT_TIMEOUT = 30        # connect() that hasn't resolved by here ⇒ wedged transport
-QUERY_TIMEOUT = 60          # sending the request is near-instant; bound it anyway
-DISCONNECT_TIMEOUT = 10     # don't let a stuck disconnect hang shutdown/reconnect
-RECV_IDLE_TIMEOUT = 300     # no stream activity for this long ⇒ treat the transport as dead
-                            # (generous: a long-running tool can legitimately go quiet a while)
-MAX_INLINE_IMAGE_BYTES = 16 * 1024 * 1024   # never base64-inline a local file bigger than this
-                            # (a multi-GB file with an image extension would otherwise be read
-                            # whole into RAM and explode the query payload)
-MAX_CHAT_LINES = 4000       # cap the rendered transcript; prune oldest lines past this so a
-                            # very long session doesn't slow Tk layout / leak embedded canvases
-MAX_CHAT_CHARS = 350_000    # also cap by characters — one giant whitespace-free assistant
-                            # line counts as 1 line and would otherwise bypass the line cap
-IMAGE_EXTS = (".png", ".jpg", ".jpeg", ".gif", ".bmp", ".webp")
-TOOL_IDLE_TIMEOUT = 1800    # once a tool call is in flight, allow a much longer silent gap
-                            # (a long build/test can legitimately stream nothing for minutes)
-COMPACT_IDLE_TIMEOUT = 600  # /compact is one big summarization round-trip that streams nothing
-                            # for a while (≈30s even on a small context); bound it generously
-MAX_PASTE_SOURCES = 8       # cap how many files one paste fans out into
-MAX_PENDING_IMAGES = 16     # cap total queued attachments (a hostile clipboard can't pile up)
-MAX_PASTE_PIXELS = 32_000_000   # reject a pasted image above this pixel count BEFORE decode/
-                            # thumbnail — a "decompression bomb" PNG decodes to a huge bitmap
-                            # (Pillow only *warns*, doesn't raise, below ~178M px)
-MAX_INLINE_IMAGES = 16      # cap images per turn (count) ...
-MAX_INLINE_TOTAL_BYTES = 32 * 1024 * 1024   # ... and aggregate bytes (the per-file cap alone
-                            # doesn't bound many-attachment memory blow-up)
-MAX_UPDATE_BODY = 1 * 1024 * 1024   # cap the update-check response body before json.loads
-MAX_UPDATE_TAGS = 300       # and the number of tags parsed
-
-# ── debug / activity log (monitoring) ──────────────────────────────────────
-# Under pythonw the overlay has no console and exposes no IPC, so its work can't be
-# watched from outside. Opt in by setting the CLAUDE_OVERLAY_DEBUG_LOG environment
-# variable to a file path: you then get a timestamped, one-line-per-event trace of the
-# worker (turn start, tool calls, results, errors, reconnects, a throttled streaming
-# heartbeat) — enough to see what it's doing and whether a turn is stuck. Default is OFF
-# (empty) so nothing is written. Privacy note: even when enabled, reply text is NEVER
-# logged (deltas/thinking are logged only as a ~2s heartbeat + char count), but a
-# turn-start prompt preview (≤120 chars) IS written — so only enable it on a trusted
-# machine. Each PID tags its own lines so several overlays don't get confused.
-DEBUG_LOG = os.environ.get("CLAUDE_OVERLAY_DEBUG_LOG", "")
-DEBUG_LOG_MAX_BYTES = 2_000_000     # truncate (best-effort) once the log grows past this
-_dbg_lock = threading.Lock()
-_dbg_stream_last = [0.0]             # throttle high-frequency streaming deltas to a heartbeat
-_dbg_think_last = [0.0]             # ditto for thinking deltas (separate, so a thinking heartbeat
-                                    # can't suppress the first answer delta we use to time TTFT)
-
-
-def dbg(kind, payload=None):
-    """Append one best-effort line to DEBUG_LOG. Never raises into the caller."""
-    if not DEBUG_LOG:
-        return
-    try:
-        if kind in ("delta", "think"):   # streaming token text → heartbeat only, NEVER the content
-            now = time.monotonic()       # (thinking is reply content too — must not hit disk verbatim)
-            last = _dbg_stream_last if kind == "delta" else _dbg_think_last
-            if now - last[0] < 2.0:
-                return
-            last[0] = now
-            n = len(payload) if isinstance(payload, str) else 0
-            payload = f"<{'streaming' if kind == 'delta' else 'thinking'} +{n} chars>"
-        elif kind == "tool":        # (name, input_dict) → name + a short arg preview
-            name, inp = payload if isinstance(payload, tuple) and len(payload) == 2 else (payload, None)
-            arg = ""
-            if isinstance(inp, dict):
-                arg = " ".join(f"{k}={str(v)[:40]}" for k, v in list(inp.items())[:2])
-            payload = f"{name} {arg}".strip()
-        elif isinstance(payload, dict):
-            payload = " ".join(f"{k}={v}" for k, v in payload.items())
-        elif isinstance(payload, str):
-            payload = payload[:200].replace("\n", " ")
-        _now = time.time()
-        _ts = time.strftime('%H:%M:%S', time.localtime(_now)) + f".{int((_now % 1) * 1000):03d}"
-        line = f"{_ts} pid={os.getpid()} {kind} {payload if payload is not None else ''}".rstrip() + "\n"
-        with _dbg_lock:
-            try:
-                if os.path.exists(DEBUG_LOG) and os.path.getsize(DEBUG_LOG) > DEBUG_LOG_MAX_BYTES:
-                    open(DEBUG_LOG, "w", encoding="utf-8").close()
-            except Exception:
-                pass
-            with open(DEBUG_LOG, "a", encoding="utf-8") as f:
-                f.write(line)
-    except Exception:
-        pass
-
-
-class _UIQueueTap:
-    """Wraps the worker→UI queue so every event the worker emits is also written to the
-    debug log. The worker's entire observable behavior already flows through this one
-    queue (`ready`/`tool`/`result`/`error`/`system`/`turn_done`/`delta`/…), so a single
-    tap here logs all of it without scattering calls through the worker. The UI keeps
-    reading the underlying queue directly; only the worker's `put` side is wrapped."""
-    def __init__(self, q):
-        self._q = q
-
-    def put(self, item, *a, **k):
-        try:
-            if isinstance(item, tuple) and len(item) == 2:
-                dbg(item[0], item[1])
-            else:
-                dbg(item)
-        except Exception:
-            pass
-        return self._q.put(item, *a, **k)
-
-    def __getattr__(self, name):
-        return getattr(self._q, name)
-
-
-SYSTEM_APPEND = (
-    "You are running as an always-on-top floating overlay assistant on the user's "
-    "Windows 11 desktop. The user talks to you without leaving their current app. "
-    "Messages may include live screenshots of the user's screen — attached directly "
-    "as images, or (legacy) as an [ATTACHMENTS] path you open with the Read tool. "
-    "Use them to see what the user is looking at, then help. "
-    "Keep replies concise and skimmable since they render in a small floating window; "
-    "expand only when asked. "
-    "When automating Office (PowerPoint/Excel/Word) via PowerShell+COM, optimize for "
-    "speed: a NEW PowerShell process runs per tool call and COM state does NOT persist "
-    "across calls, and every property access is a slow cross-process round-trip. So: "
-    "(1) BATCH — do all inspection in ONE script (return what you need, e.g. as JSON), "
-    "then apply ALL edits in ONE script; never one tool call per shape/cell/slide. "
-    "(2) Within a script cache COM references in variables (grab the slide/shape/table "
-    "once) instead of re-walking the object model, and don't re-read everything to verify "
-    "after each write. (3) For Excel bulk writes, set Application.ScreenUpdating=$false, "
-    "Calculation=xlManual and EnableEvents=$false around them, then restore. (4) For large "
-    "purely-textual edits where the live open document isn't needed, python-pptx/openpyxl "
-    "on the file is far faster than COM — but only when the file is NOT open in Office."
-)
-
-THEMES = {
-    "light": {
-        "bg": "#FAF9F5", "field": "#FFFFFF", "user_card": "#EFEBE1",
-        "text": "#28261F", "muted": "#73706A", "faint": "#A9A59B",
-        "accent": "#D97757", "accent_hi": "#C25E40", "on_accent": "#FFFFFF",
-        "border": "#E6E2D8", "tool_bg": "#F2EFE7", "err": "#B4413A",
-        "sel": "#EADDD3", "hover": "#EFEBE1",
-    },
-    "dark": {
-        "bg": "#262624", "field": "#1F1E1D", "user_card": "#34332F",
-        "text": "#ECEAE3", "muted": "#9B978D", "faint": "#6F6C64",
-        "accent": "#D97757", "accent_hi": "#E68A6C", "on_accent": "#FFFFFF",
-        "border": "#3A3934", "tool_bg": "#2E2D2A", "err": "#E0897D",
-        "sel": "#3A3934", "hover": "#30302E",
-    },
-}
-T = THEMES.get(THEME, THEMES["light"])
-
-
-def set_dpi_awareness():
-    """Make the process DPI-aware so 1 Tk pixel == 1 physical pixel (crisp, no
-    OS bitmap-stretch). Must run before the Tk interpreter starts."""
-    try:
-        ctypes.windll.shcore.SetProcessDpiAwareness(2)   # PER-MONITOR aware
-    except Exception:
-        try:
-            ctypes.windll.user32.SetProcessDPIAware()
-        except Exception:
-            pass
-
-
-def set_app_user_model_id():
-    """Give the process an explicit AppUserModelID so the Windows taskbar shows OUR
-    window icon (APP_ICON) rather than pythonw's, and doesn't lump the overlay together
-    with other Python apps. Must run before the first window is created."""
-    if not (TASKBAR_BUTTON and APP_ID):
-        return
-    try:
-        ctypes.windll.shell32.SetCurrentProcessExplicitAppUserModelID(APP_ID)
-    except Exception:
-        pass
-
-
-# Win32 region calls — set argtypes so 64-bit handles aren't truncated.
-_gdi32, _user32 = ctypes.windll.gdi32, ctypes.windll.user32
-_gdi32.CreateRoundRectRgn.restype = wt.HRGN
-_gdi32.CreateRoundRectRgn.argtypes = [ctypes.c_int] * 6
-_user32.SetWindowRgn.restype = ctypes.c_int
-_user32.SetWindowRgn.argtypes = [wt.HWND, wt.HRGN, ctypes.c_bool]
-_gdi32.DeleteObject.restype = ctypes.c_int
-_gdi32.DeleteObject.argtypes = [ctypes.c_void_p]   # free a region Windows didn't take ownership of
-_user32.GetAncestor.restype = wt.HWND
-_user32.GetAncestor.argtypes = [wt.HWND, ctypes.c_uint]
-_gdi32.CreateEllipticRgn.restype = wt.HRGN
-_gdi32.CreateEllipticRgn.argtypes = [ctypes.c_int] * 4
-# Region from arbitrary silhouette (used to float the collapsed orb as a pixel sprite,
-# with no circular frame): OR together one rect per opaque run of the artwork's alpha.
-_gdi32.CreateRectRgn.restype = wt.HRGN
-_gdi32.CreateRectRgn.argtypes = [ctypes.c_int] * 4
-_gdi32.SetRectRgn.restype = ctypes.c_int
-_gdi32.SetRectRgn.argtypes = [wt.HRGN] + [ctypes.c_int] * 4
-_gdi32.CombineRgn.restype = ctypes.c_int
-_gdi32.CombineRgn.argtypes = [wt.HRGN, wt.HRGN, wt.HRGN, ctypes.c_int]
-_user32.GetMonitorInfoW.argtypes = [ctypes.c_void_p, ctypes.c_void_p]
-_user32.GetMonitorInfoW.restype = ctypes.c_int
-_MONENUMPROC = ctypes.WINFUNCTYPE(ctypes.c_int, ctypes.c_void_p, ctypes.c_void_p,
-                                  ctypes.POINTER(wt.RECT), ctypes.c_void_p)
-_user32.EnumDisplayMonitors.argtypes = [ctypes.c_void_p, ctypes.c_void_p, _MONENUMPROC, ctypes.c_void_p]
-_user32.EnumDisplayMonitors.restype = ctypes.c_int
-_user32.SetWindowDisplayAffinity.argtypes = [wt.HWND, ctypes.c_uint]
-_user32.SetWindowDisplayAffinity.restype = ctypes.c_int
-_user32.GetForegroundWindow.restype = wt.HWND
-# Extended window-style get/set — used to force a taskbar button onto the frameless
-# (overrideredirect) window. argtypes set so the 64-bit HWND isn't truncated.
-_user32.GetWindowLongW.restype = ctypes.c_long
-_user32.GetWindowLongW.argtypes = [wt.HWND, ctypes.c_int]
-_user32.SetWindowLongW.restype = ctypes.c_long
-_user32.SetWindowLongW.argtypes = [wt.HWND, ctypes.c_int, ctypes.c_long]
-GWL_EXSTYLE      = -20
-WS_EX_TOOLWINDOW = 0x00000080   # no taskbar button (what an overrideredirect popup effectively is)
-WS_EX_APPWINDOW  = 0x00040000   # force a taskbar button even on a tool/popup window
-# Z-order raise + forced activation — used when the taskbar button / alt-tab / hotkey asks for
-# the overlay: the OS activates it but does NOT re-order it above other always-on-top windows,
-# so a click could leave it buried under another topmost window (or just unfocused). These do
-# the raise the activation skips. Plain z-order, no SetWindowRgn → clear of the v1.1.9 freeze class.
-HWND_TOPMOST     = wt.HWND(-1)
-SWP_NOSIZE       = 0x0001
-SWP_NOMOVE       = 0x0002
-SWP_NOACTIVATE   = 0x0010
-SWP_SHOWWINDOW   = 0x0040
-_user32.SetWindowPos.argtypes = [wt.HWND, wt.HWND, ctypes.c_int, ctypes.c_int,
-                                 ctypes.c_int, ctypes.c_int, ctypes.c_uint]
-_user32.SetWindowPos.restype = wt.BOOL
-_user32.BringWindowToTop.argtypes = [wt.HWND]
-_user32.BringWindowToTop.restype = wt.BOOL
-_user32.SetForegroundWindow.argtypes = [wt.HWND]
-_user32.SetForegroundWindow.restype = wt.BOOL
-_user32.GetWindowThreadProcessId.argtypes = [wt.HWND, ctypes.POINTER(wt.DWORD)]
-_user32.GetWindowThreadProcessId.restype = wt.DWORD
-_user32.AttachThreadInput.argtypes = [wt.DWORD, wt.DWORD, wt.BOOL]
-_user32.AttachThreadInput.restype = wt.BOOL
-_user32.IsClipboardFormatAvailable.argtypes = [ctypes.c_uint]
-_user32.IsClipboardFormatAvailable.restype = ctypes.c_int
-# Standard clipboard format ids — used for a cheap, non-blocking "is there an image?" probe
-# on the UI thread, so we only spin up the (potentially slow) ImageGrab.grabclipboard() read
-# on a background thread when there's actually image/file content.
-CF_BITMAP, CF_DIB, CF_HDROP, CF_DIBV5 = 2, 8, 15, 17
-CF_TEXT, CF_UNICODETEXT = 1, 13   # so a text copy (which many apps ALSO put a bitmap on the
-                                  # clipboard for) pastes as text, not as an image
-
-# Exclude the overlay from screen captures at the OS level (DWM): the window stays
-# visible to the user but is omitted from PIL ImageGrab / PrintWindow, so the
-# screenshots we send Claude never contain the overlay obscuring the content — and
-# we no longer have to withdraw() + sleep() on every capture. Verified on this
-# machine (returns the content behind the window, not black).
-# WDA_NONE clears the affinity again → the window becomes capturable (shows up in
-# Teams/Zoom/OBS screen shares); the "shareable" status-bar toggle flips between them.
-WDA_NONE = 0x00
-WDA_EXCLUDEFROMCAPTURE = 0x11
-
-
-class _MONITORINFO(ctypes.Structure):
-    _fields_ = [("cbSize", wt.DWORD), ("rcMonitor", wt.RECT),
-                ("rcWork", wt.RECT), ("dwFlags", wt.DWORD)]
-
-
-def enumerate_monitors():
-    """Return [{'rect': (l, t, r, b), 'primary': bool}, ...], primary first."""
-    mons = []
-
-    def _cb(hmon, hdc, lprc, lparam):
-        mi = _MONITORINFO()
-        mi.cbSize = ctypes.sizeof(_MONITORINFO)
-        if _user32.GetMonitorInfoW(hmon, ctypes.byref(mi)):
-            r = mi.rcMonitor
-            mons.append({"rect": (r.left, r.top, r.right, r.bottom),
-                         "primary": bool(mi.dwFlags & 1)})   # MONITORINFOF_PRIMARY
-        return 1
-
-    try:
-        proc = _MONENUMPROC(_cb)
-        _user32.EnumDisplayMonitors(None, None, proc, 0)
-    except Exception:
-        pass
-    mons.sort(key=lambda m: (not m["primary"], m["rect"][0], m["rect"][1]))  # primary, then L→R
-    return mons
-
-
-# ───────────────────────── background Claude worker ───────────────────────
-class ClaudeWorker(threading.Thread):
-    def __init__(self, ui_queue: "queue.Queue"):
-        super().__init__(daemon=True)
-        # Tap the UI channel so the debug log captures every worker→UI event (no-op when
-        # DEBUG_LOG is ""). The UI side keeps reading the raw queue.
-        self.ui = _UIQueueTap(ui_queue) if DEBUG_LOG else ui_queue
-        self.req: "queue.Queue" = queue.Queue()
-        self._loop: asyncio.AbstractEventLoop | None = None
-        self._client: ClaudeSDKClient | None = None
-        self._running = True
-        self._saw_stream = False
-        self._lifecycle_task = None   # the in-flight connect()/disconnect() task, if any
-
-    def ask(self, text: str, image_paths=None):
-        self.req.put(("ask", (text, list(image_paths or []))))
-    def reset(self):                  self.req.put(("reset", None))
-    def compact(self):                self.req.put(("compact", None))
-    def shutdown(self):
-        self._running = False
-        # If the worker is currently AWAITING a lifecycle call (connect/disconnect), the
-        # queued "stop" can't be read until that await returns (up to CONNECT_TIMEOUT).
-        # Cancel the in-flight lifecycle task so the worker can wind down promptly instead
-        # of leaving a daemon thread + orphaned `claude` CLI child after the UI is gone.
-        loop, task = self._loop, self._lifecycle_task
-        if loop and task and not loop.is_closed():
-            try:
-                loop.call_soon_threadsafe(task.cancel)
-            except Exception:
-                pass
-        self.req.put(("stop", None))
-
-    def interrupt(self):
-        loop, client = self._loop, self._client
-        # The loop may be closed (worker finished / between restarts) — calling
-        # run_coroutine_threadsafe on a closed loop raises RuntimeError straight into the
-        # Tk callback (reset()/Stop don't guard it) and leaks the coroutine object.
-        if not (loop and client) or loop.is_closed():
-            return
-        coro = self._safe_interrupt(client)
-        try:
-            asyncio.run_coroutine_threadsafe(coro, loop)
-        except RuntimeError:
-            try:
-                coro.close()
-            except Exception:
-                pass
-
-    async def _safe_interrupt(self, client):
-        try:
-            await client.interrupt()
-        except Exception:
-            pass
-
-    def set_model(self, model):
-        # Go through the request queue (not run_coroutine_threadsafe) so a model switch is
-        # serialized behind any queued reset/ask and can't interleave with _close() tearing
-        # down the same client — which could leave a half-disconnected client or a status
-        # line stuck on "switching model…".
-        self.req.put(("set_model", model))
-
-    async def _do_set_model(self, client, model):
-        try:
-            await client.set_model(model)
-            await self._emit_usage()
-            self.ui.put(("status", ""))   # clear the "switching model…" notice
-        except Exception as e:
-            self.ui.put(("error", f"set_model failed: {type(e).__name__}: {e}"))
-
-    async def _allow_tool(self, tool_name, input_data, context):
-        # Auto-approve every tool. permission_mode="bypassPermissions" already does
-        # this on most machines, but managed/enterprise installs can DISABLE bypass
-        # mode (managed-settings.json: disableBypassPermissionsMode), which makes the
-        # CLI fall back to "default" and emit a permission prompt. The overlay is a
-        # GUI with no TTY, so an unanswered prompt would just hang the turn forever
-        # ("nowhere to approve"). This callback answers those prompts so the overlay
-        # works regardless of the host's permission policy. The tool call still shows
-        # up as a chip in the chat via the normal streaming path, so it isn't silent.
-        return PermissionResultAllow()
-
-    def _make_options(self) -> ClaudeAgentOptions:
-        opts = dict(
-            permission_mode=PERMISSION_MODE, cwd=WORKING_DIR, model=MODEL,
-            can_use_tool=self._allow_tool,
-            include_partial_messages=True,
-            # exclude_dynamic_sections strips the per-turn-changing bits (cwd, git
-            # status, auto-memory) out of the preset system prompt so the big static
-            # prefix stays byte-stable → prompt-cache hits survive across turns.
-            system_prompt={"type": "preset", "preset": "claude_code",
-                           "append": SYSTEM_APPEND, "exclude_dynamic_sections": True},
-        )
-        opts["max_buffer_size"] = MAX_BUFFER_SIZE
-        if SKILLS is not None:
-            # Enable Agent SDK skills (e.g. the slide-check skill). The SDK only injects the
-            # Skill tool + sets setting_sources=["user","project"] when this is provided; left
-            # unset the overlay discovers no skills at all. A list enables only those skills.
-            opts["skills"] = SKILLS
-        if STRICT_MCP_CONFIG:
-            # Use ONLY the (empty) MCP servers defined here, ignoring the user's filesystem
-            # config. Without this the spawned CLI loads every MCP server from ~/.claude.json
-            # and injects all their tool schemas — measured at 72K tokens (36% of Haiku's
-            # 200K window) on one machine with many MCP servers, gone before the first
-            # message. setting_sources
-            # alone does NOT stop this; the CLI loads MCP servers via a separate path.
-            opts["strict_mcp_config"] = True
-        # Some kwargs (max_buffer_size, can_use_tool, strict_mcp_config) only exist on newer
-        # SDKs. Strip any the installed SDK rejects, one at a time, so an older install still
-        # loads (with reduced features) instead of failing to construct options at all.
-        droppable = ["strict_mcp_config", "max_buffer_size", "can_use_tool",
-                     "include_partial_messages", "skills"]
-        while True:
-            try:
-                return ClaudeAgentOptions(**opts)
-            except TypeError as e:
-                victim = next((k for k in droppable if k in opts and k in str(e)), None)
-                if victim is None:
-                    victim = next((k for k in droppable if k in opts), None)
-                if victim is None:
-                    raise
-                opts.pop(victim, None)
-
-    def run(self):
-        # Bounded auto-restart: even if _amain falls over entirely (e.g. the event loop
-        # dies), bring it back so the overlay self-heals instead of becoming a zombie
-        # window that never answers again.
-        dbg("worker_start")
-        attempts = 0
-        last_start = 0.0
-        while self._running and attempts < 5:
-            now = time.monotonic()
-            if last_start and now - last_start > 180:
-                attempts = 0           # survived a stable stretch → forget old failures, so
-                                       # rare crashes spread over a long session don't add up
-                                       # to a permanent "stopped" state (storm-based, not lifetime)
-            last_start = now
-            attempts += 1
-            try:
-                asyncio.run(self._amain())
-                return                      # _amain returned cleanly (stop requested)
-            except BaseException as e:  # pragma: no cover  (BaseException: e.g. CancelledError)
-                self.ui.put(("error", f"worker restarting after: {type(e).__name__}: {e}"))
-                self._client = None
-                time.sleep(0.5)
-            finally:
-                # asyncio.run() closed this loop; null it so interrupt()/set_model() don't
-                # schedule onto a dead loop before the next iteration sets a fresh one.
-                self._loop = None
-        if self._running:
-            self.ui.put(("error", "Claude worker stopped after repeated failures — "
-                                  "please restart the overlay."))
-
-    async def _amain(self):
-        self._loop = asyncio.get_running_loop()
-        await self._open()
-        while self._running:
-            try:
-                kind, payload = await self._loop.run_in_executor(None, self.req.get)
-            except Exception:
-                continue
-            if kind == "stop":
-                break
-            # Each request is fully guarded: a failure here must never break the loop
-            # (that would leave the UI waiting on a worker that's gone). Worst case we
-            # reconnect and keep serving.
-            try:
-                if kind == "reset":
-                    await self._close()
-                    self._saw_stream = False
-                    await self._open()
-                    self.ui.put(("reset_done", None))
-                elif kind == "ask":
-                    await self._run_turn(payload)
-                elif kind == "compact":
-                    await self._run_compact()
-                elif kind == "set_model":
-                    if self._client is None:
-                        self.ui.put(("error", "Not connected to Claude yet — can't switch model."))
-                        self.ui.put(("status", ""))
-                    else:
-                        await self._do_set_model(self._client, payload)
-            except asyncio.CancelledError:
-                # a cancel (Stop / transport teardown) must not break the loop or be
-                # mistaken for a fatal error — CancelledError is BaseException, not
-                # Exception, so it would otherwise escape and kill the worker.
-                self.ui.put(("turn_done", None))
-            except BaseException as e:
-                self.ui.put(("error", f"{type(e).__name__}: {e}"))
-                self.ui.put(("turn_done", None))
-                await self._reconnect()
-        await self._close()
-
-    async def _reconnect(self):
-        """Tear down a broken client and stand up a fresh one so the next turn works.
-        The conversation context is lost (new session), but the app stays alive instead
-        of freezing on a dead transport."""
-        self.ui.put(("system", "↻ Connection hiccup — reconnected with a fresh session."))
-        try:
-            await self._close()
-        except Exception:
-            pass
-        self._saw_stream = False
-        await self._open()
-
-    async def _open(self):
-        try:
-            self._client = ClaudeSDKClient(options=self._make_options())
-            # Bound the connect: a wedged transport (TLS MITM, half-open socket, CLI stuck on
-            # a prompt) would otherwise hang the worker here forever, where no reconnect/restart
-            # guard can reach it. A timeout degrades to the normal "couldn't start" path.
-            # Run it as a tracked task so shutdown() can cancel it (see shutdown/_lifecycle_task).
-            self._lifecycle_task = asyncio.ensure_future(self._client.connect())
-            try:
-                await asyncio.wait_for(self._lifecycle_task, CONNECT_TIMEOUT)
-            finally:
-                self._lifecycle_task = None
-            self.ui.put(("ready", None))
-            # Context usage is informative only; do not block the first queued prompt on
-            # an extra CLI round-trip during startup/reconnect.
-            try:
-                self._loop.create_task(self._emit_usage())
-            except Exception:
-                pass
-        except BaseException as e:   # incl. CancelledError — _open must never propagate
-            self._client = None
-            if isinstance(e, (asyncio.TimeoutError, TimeoutError)):
-                self.ui.put(("error",
-                    f"Connecting to Claude timed out after {CONNECT_TIMEOUT}s. The next "
-                    "message will try again. (Check your network / `claude --version`.)"))
-            elif isinstance(e, TypeError):   # ClaudeAgentOptions rejected a kwarg → SDK too old
-                self.ui.put(("error",
-                    f"Your claude-agent-sdk looks too old ({type(e).__name__}: {e}). "
-                    "Update it:  pip install --upgrade claude-agent-sdk  (or run update.cmd)."))
-            else:
-                self.ui.put(("error",
-                    f"Could not start Claude: {type(e).__name__}: {e}\n"
-                    "Is the `claude` CLI installed and logged in? Run `claude --version` "
-                    "in a terminal; if it's missing, run setup.cmd (or `irm "
-                    "https://claude.ai/install.ps1 | iex`), then `claude` to /login."))
-
-    async def _emit_usage(self):
-        """Push current model + context-window usage % to the UI statusline."""
-        # Capture the client we're measuring. A turn's finally schedules this against the
-        # *current* client; if a Clear/reconnect swaps the client out while the (slow,
-        # round-trips to the CLI) get_context_usage() is in flight, the result describes a
-        # session that no longer exists. Emitting it would overwrite the fresh post-reset
-        # baseline with the OLD conversation's high % — the "Clear didn't drop context" bug.
-        client = self._client
-        if client is None:
-            return
-        try:
-            u = await asyncio.wait_for(client.get_context_usage(), timeout=6)
-            if client is not self._client:   # reset/reconnect happened mid-flight → stale
-                return
-            if isinstance(u, dict):
-                if u.get("model"):
-                    self.ui.put(("model", u["model"]))
-                if u.get("percentage") is not None:
-                    self.ui.put(("ctx", u["percentage"]))
-        except Exception:
-            pass
-
-    async def _close(self):
-        # Null the handle FIRST so a disconnect that hangs (bounded below) can't leave the
-        # rest of the worker pointing at a half-dead client.
-        client, self._client = self._client, None
-        if client is not None:
-            self._lifecycle_task = asyncio.ensure_future(client.disconnect())
-            try:
-                await asyncio.wait_for(self._lifecycle_task, DISCONNECT_TIMEOUT)
-            except Exception:
-                pass
-            finally:
-                self._lifecycle_task = None
-
-    async def _run_turn(self, payload):
-        text, image_paths = payload if isinstance(payload, tuple) else (payload, [])
-        dbg("turn_start", f"imgs={len(image_paths or [])} | {str(text)[:120]}")
-        _dbg_stream_last[0] = 0.0   # force this turn's FIRST delta to log → measures time-to-first-token
-        _dbg_think_last[0] = 0.0    # and its first thinking token → time-to-first-thinking
-        # ── Debug-only test hook, gated by CLAUDE_OVERLAY_DEBUG_LOG so it's inert in normal use
-        #    (without the env var, `/simerror` is just sent to Claude as ordinary text). It emits a
-        #    synthetic errored ResultMessage so the "last turn ended with an error (<reason>)" UI
-        #    can be exercised without a real API failure. Presets: (none)=overloaded, max, exec,
-        #    rate; any other word is used verbatim as the subtype.
-        ts = text.strip() if isinstance(text, str) else ""
-        if DEBUG_LOG and (ts == "/simerror" or ts.startswith("/simerror ")):
-            arg = ts[len("/simerror"):].strip()
-            presets = {
-                "":     ("overloaded_error", "The model was overloaded (HTTP 529). Transient — the next turn retries."),
-                "max":  ("error_max_turns", None),
-                "exec": ("error_during_execution", "A tool call failed during execution."),
-                "rate": ("rate_limit_error", "Rate limited (HTTP 429)."),
-            }
-            sub, detail = presets.get(arg, (arg, None))
-            self.ui.put(("system", f"(test: simulating an errored result — subtype={sub})"))
-            self.ui.put(("result", {"is_error": True, "subtype": sub, "result": detail,
-                                    "stop_reason": None, "cost": None}))
-            self.ui.put(("turn_done", None))
-            return
-        if self._client is None:        # initial connect failed earlier — try once more
-            await self._open()
-        if self._client is None:
-            self.ui.put(("error", "Not connected to Claude. Check `claude --version`."))
-            self.ui.put(("turn_done", None))
-            return
-        agen = None
-        try:
-            await asyncio.wait_for(
-                self._client.query(self._build_query(text, image_paths)), QUERY_TIMEOUT)
-            blocks: dict = {}
-            tool_active = False
-            # Iterate the stream item-by-item under an idle timeout instead of a bare
-            # `async for`: if the transport goes silent forever (dead CLI, wedged socket)
-            # the turn would otherwise hold "thinking…" indefinitely. A gap longer than the
-            # idle budget is treated as a broken transport → reconnect. Once a tool call is in
-            # flight we switch to a much longer budget so a legitimately silent long-running
-            # tool (a big build/test that streams nothing for minutes) isn't mistaken for dead.
-            agen = self._client.receive_response()
-            while True:
-                budget = TOOL_IDLE_TIMEOUT if tool_active else RECV_IDLE_TIMEOUT
-                try:
-                    msg = await asyncio.wait_for(agen.__anext__(), budget)
-                except StopAsyncIteration:
-                    break
-                if not tool_active and self._msg_has_tool(msg):
-                    tool_active = True
-                self._dispatch(msg, blocks)
-        except asyncio.CancelledError:
-            # Stop button / interrupt() / transport cancel — end this turn cleanly.
-            # (BaseException, so it'd otherwise escape every `except Exception` and the
-            # worker thread would die permanently.) Don't reconnect; shutdown is queue-driven.
-            self.ui.put(("system", "⏹ stopped."))
-        except (asyncio.TimeoutError, TimeoutError):
-            # query() wedged or the stream went silent past the idle budget → the transport
-            # is effectively dead; rebuild it so the next turn works instead of hanging here.
-            self.ui.put(("error", "Claude stopped responding — reconnecting with a fresh session."))
-            await self._reconnect()
-        except BaseException as e:
-            self.ui.put(("error", f"{type(e).__name__}: {e}"))
-            # a decode/connection/process error means the transport is dead — the client
-            # is unusable now, so rebuild it before the next turn instead of erroring
-            # forever (the classic "it crashed and won't respond anymore" symptom).
-            if isinstance(e, (CLIJSONDecodeError, CLIConnectionError, ProcessError, ClaudeSDKError)):
-                await self._reconnect()
-        finally:
-            # Finalize the response stream. wait_for cancelling __anext__() does NOT close the
-            # async generator, so without this the SDK's reader task / stdout pipe can be left
-            # half-open (a leak, or a later disconnect() that hangs). Bounded so a broken close
-            # can't reintroduce a hang.
-            if agen is not None:
-                aclose = getattr(agen, "aclose", None)
-                if aclose is not None:
-                    try:
-                        await asyncio.wait_for(aclose(), DISCONNECT_TIMEOUT)
-                    except BaseException:
-                        pass
-            self.ui.put(("turn_done", None))
-            # Refresh context% off the critical path: schedule it rather than
-            # awaiting, so the UI leaves "thinking…" the instant the reply ends
-            # instead of after an extra round-trip.
-            try:
-                self._loop.create_task(self._emit_usage())
-            except Exception:
-                pass
-
-    async def _run_compact(self):
-        """Compact the conversation by sending the CLI's `/compact` command. Compaction
-        streams only system/user messages (status → compact_boundary → the re-injected
-        summary → an empty result) — no assistant answer — so we DON'T render the stream as
-        chat. We just signal start/finish to the UI (which animates a CLI-style spinner) and
-        capture the compact_boundary's pre/post token counts for the result line."""
-        if self._client is None:        # not connected yet — try once
-            await self._open()
-        if self._client is None:
-            self.ui.put(("compact_done",
-                         {"status": "error", "meta": None,
-                          "detail": "not connected (check `claude --version`)"}))
-            return
-        self.ui.put(("compacting", None))   # → UI starts the animation
-        agen = None
-        meta = None
-        status = "ok"
-        detail = None
-        result_signal = None                # the CLI's own compact_result flag, if it sends one
-        try:
-            await asyncio.wait_for(self._client.query("/compact"), QUERY_TIMEOUT)
-            agen = self._client.receive_response()
-            while True:
-                try:
-                    msg = await asyncio.wait_for(agen.__anext__(), COMPACT_IDLE_TIMEOUT)
-                except StopAsyncIteration:
-                    break
-                info = self._compact_meta(msg)
-                if info is not None:
-                    meta = info
-                sig = self._compact_result_signal(msg)
-                if sig is not None:
-                    result_signal = sig
-            # The stream finished cleanly — but "no exception" is NOT "it compacted". Confirm it:
-            # trust the CLI's compact_result=="success" flag, OR a compact_boundary that shows the
-            # context actually shrank (post < pre). Otherwise report "unconfirmed" rather than
-            # claiming success (e.g. the CLI declined / nothing to compact / a silent internal fail).
-            shrank = bool(meta and meta.get("pre_tokens") and meta.get("post_tokens")
-                          and meta["post_tokens"] < meta["pre_tokens"])
-            if result_signal == "success" or shrank:
-                status = "ok"
-            else:
-                status = "unconfirmed"
-                detail = (f"CLI reported compact_result={result_signal!r}"
-                          if result_signal is not None
-                          else "no success signal or token reduction seen")
-        except asyncio.CancelledError:
-            status = "cancelled"        # Stop / Clear interrupted it — conversation unchanged
-        except (asyncio.TimeoutError, TimeoutError):
-            status = "timeout"
-            await self._reconnect()
-        except BaseException as e:
-            status = "error"
-            detail = f"{type(e).__name__}: {e}"
-            if isinstance(e, (CLIJSONDecodeError, CLIConnectionError, ProcessError, ClaudeSDKError)):
-                await self._reconnect()
-        finally:
-            if agen is not None:
-                aclose = getattr(agen, "aclose", None)
-                if aclose is not None:
-                    try:
-                        await asyncio.wait_for(aclose(), DISCONNECT_TIMEOUT)
-                    except BaseException:
-                        pass
-            self.ui.put(("compact_done", {"status": status, "meta": meta, "detail": detail}))
-            try:
-                self._loop.create_task(self._emit_usage())
-            except Exception:
-                pass
-
-    @staticmethod
-    def _compact_meta(msg):
-        """Pull {pre_tokens, post_tokens, duration_ms, trigger} from a compact_boundary
-        system message (the authoritative before/after sizes), else None. Identified by
-        class name so this doesn't depend on importing the SDK's message types."""
-        if type(msg).__name__ != "SystemMessage":
-            return None
-        if getattr(msg, "subtype", None) != "compact_boundary":
-            return None
-        data = getattr(msg, "data", None)
-        md = data.get("compact_metadata") if isinstance(data, dict) else None
-        if not isinstance(md, dict):
-            return None
-        return {"pre_tokens": md.get("pre_tokens"), "post_tokens": md.get("post_tokens"),
-                "duration_ms": md.get("duration_ms"), "trigger": md.get("trigger")}
-
-    @staticmethod
-    def _compact_result_signal(msg):
-        """The CLI's own success flag: a status system message carries
-        `compact_result: "success"` (or an error string) once compaction settles.
-        Returns that value, or None for messages that don't carry it."""
-        if type(msg).__name__ != "SystemMessage":
-            return None
-        data = getattr(msg, "data", None)
-        if not isinstance(data, dict):
-            return None
-        cr = data.get("compact_result")
-        return cr if cr else None
-
-    def _build_query(self, text: str, image_paths: list):
-        """Return the prompt for client.query(). With inline images we yield a
-        structured user message (text + base64 image blocks) so the model sees
-        the screen directly — no per-turn Read round-trip. Otherwise a plain
-        string (the legacy "Read the PNG path" flow builds its own text upstream)."""
-        if IMAGE_INPUT != "inline" or not image_paths:
-            return text
-        content: list = []
-        if text:
-            content.append({"type": "text", "text": text})
-        failed = 0
-        total = 0
-        seen = set()
-        for p in image_paths:
-            if p in seen:           # dedupe repeated paths (same screenshot/paste twice)
-                continue
-            seen.add(p)
-            if len(seen) > MAX_INLINE_IMAGES:   # cap count per turn
-                failed += 1
-                continue
-            try:
-                # Cap before reading: per-file AND aggregate, so a huge non-image file (per
-                # file) or many accumulated attachments (aggregate) can't be read whole into
-                # RAM and base64-expanded into one query.
-                size = Path(p).stat().st_size
-                if size > MAX_INLINE_IMAGE_BYTES or (total + size) > MAX_INLINE_TOTAL_BYTES:
-                    failed += 1
-                    continue
-                data = Path(p).read_bytes()
-            except Exception:
-                failed += 1
-                continue
-            if not data:            # 0-byte / unreadable-as-empty → don't send a blank block
-                failed += 1
-                continue
-            total += size
-            ext = Path(p).suffix.lower()
-            mt = {".jpg": "image/jpeg", ".jpeg": "image/jpeg", ".webp": "image/webp",
-                  ".gif": "image/gif"}.get(ext, "image/png")
-            content.append({"type": "image", "source": {
-                "type": "base64", "media_type": mt,
-                "data": base64.b64encode(data).decode()}})
-        if failed:   # tell the user their screen/image didn't actually attach
-            self.ui.put(("error", f"{failed} image(s) couldn't be read and were not sent."))
-        if not content:
-            return text
-        msg = {"type": "user",
-               "message": {"role": "user", "content": content},
-               "parent_tool_use_id": None}
-
-        async def _one():
-            yield msg
-
-        return _one()
-
-    @staticmethod
-    def _msg_has_tool(msg):
-        """True if this stream message starts/contains a tool_use — used to extend the
-        receive idle budget so a long, silent tool isn't mistaken for a dead transport."""
-        try:
-            if isinstance(msg, StreamEvent):
-                ev = msg.event or {}
-                if ev.get("type") == "content_block_start":
-                    return ((ev.get("content_block") or {}).get("type") == "tool_use")
-            elif isinstance(msg, AssistantMessage):
-                return any(isinstance(b, ToolUseBlock)
-                           for b in (getattr(msg, "content", None) or []))
-        except Exception:
-            pass
-        return False
-
-    def _dispatch(self, msg, blocks: dict):
-        # The contents are untrusted CLI stream-json — a single malformed frame
-        # (non-dict block value, unhashable index, content=None, …) must never abort
-        # the turn (which would also skip the reconnect logic). Skip the bad frame and
-        # keep streaming.
-        try:
-            self._dispatch_inner(msg, blocks)
-        except Exception:
-            pass
-
-    def _dispatch_inner(self, msg, blocks: dict):
-        if isinstance(msg, StreamEvent):
-            self._saw_stream = True
-            ev = msg.event or {}
-            t = ev.get("type")
-            if t == "content_block_start":
-                idx = ev.get("index")
-                cb = ev.get("content_block", {}) or {}
-                blocks[idx] = {"type": cb.get("type"), "name": cb.get("name"), "buf": ""}
-            elif t == "content_block_delta":
-                idx = ev.get("index")
-                d = ev.get("delta", {}) or {}
-                dt = d.get("type")
-                if dt == "text_delta":
-                    self.ui.put(("delta", d.get("text", "")))
-                elif dt == "thinking_delta":   # extended-thinking tokens (stream them so the
-                    self.ui.put(("think", d.get("thinking", "")))   # pre-answer wait looks alive
-                elif dt == "input_json_delta":
-                    b = blocks.get(idx)
-                    if not isinstance(b, dict):   # corrupted/missing → reset to a fresh buf
-                        b = {"type": "tool_use", "name": None, "buf": ""}
-                        blocks[idx] = b
-                    b["buf"] = (b.get("buf") or "") + (d.get("partial_json") or "")
-            elif t == "content_block_stop":
-                idx = ev.get("index")
-                b = blocks.get(idx)
-                if isinstance(b, dict) and b.get("type") == "tool_use":
-                    try:
-                        inp = json.loads(b.get("buf") or "{}")
-                    except Exception:
-                        inp = {}
-                    self.ui.put(("tool", (b.get("name") or "tool", inp)))
-        elif isinstance(msg, AssistantMessage):
-            if getattr(msg, "model", None):
-                self.ui.put(("model", msg.model))
-            if not self._saw_stream:
-                for blk in (getattr(msg, "content", None) or []):
-                    if isinstance(blk, TextBlock):
-                        self.ui.put(("delta", blk.text))
-                    elif isinstance(blk, ToolUseBlock):
-                        self.ui.put(("tool", (blk.name, blk.input)))
-        elif isinstance(msg, ResultMessage):
-            is_err = getattr(msg, "is_error", False)
-            subtype = getattr(msg, "subtype", None)
-            detail = getattr(msg, "result", None)
-            stop_reason = getattr(msg, "stop_reason", None)
-            if is_err:   # log the REAL reason (error detail, not reply text) so a past
-                         # occurrence is diagnosable from the activity log
-                dbg("result_error", "subtype=%s stop=%s detail=%s"
-                    % (subtype, stop_reason, str(detail)[:300]))
-            self.ui.put(("result", {"cost": getattr(msg, "total_cost_usd", None),
-                                    "is_error": is_err, "subtype": subtype,
-                                    "result": detail, "stop_reason": stop_reason}))
-
+from config import *
+from config import __version__
+from debuglog import dbg, DEBUG_LOG
+from win32utils import *
+from win32utils import _user32, _gdi32
+from worker import ClaudeWorker
 
 # ───────────────────────────── the overlay UI ─────────────────────────────
 PLACEHOLDER = "Reply to Claude…"
