+++ pytest.ini
@@ -1,0 +1,6 @@
+[pytest]
+# The modules live at the repo root (config.py, worker.py, …), so put the root on
+# sys.path for the tests in tests/ (pytest's default only adds the test dir).
+pythonpath = .
+testpaths = tests
+addopts = -q
